making a test for abstraction violation
