+++ python/ray/autoscaler/command_runner.py
@@ -105,27 +105,6 @@ class CommandRunnerInterface:
 
     Command runner instances are returned by provider.get_command_runner()."""
 
-    def __init__(self):
-        """Enforces the CommandRunnerInterface functions on the subclasses.
-
-        This is important to make sure the subclasses do not violate the
-        function abstractions. If you need to add a new function to one of
-        the CommandRunnerInterface subclasses, you have to add it to
-        CommandRunnerInterface and all of its subclasses.
-        """
-
-        subclass_available_functions = dir(self)
-        subclass_public_functions = set([
-            func for func in subclass_available_functions
-            if not func.startswith("_")
-        ])
-        cmd_runner_interface_public_functions = dir(CommandRunnerInterface)
-        allowed_public_interface_functions = set([
-            func for func in cmd_runner_interface_public_functions
-            if not func.startswith("_")
-        ])
-        assert allowed_public_interface_functions == subclass_public_functions
-
     def run(
             self,
             cmd: str = None,
@@ -202,7 +181,6 @@ class CommandRunnerInterface:
 class KubernetesCommandRunner(CommandRunnerInterface):
     def __init__(self, log_prefix, namespace, node_id, auth_config,
                  process_runner):
-        super().__init__()
         self.log_prefix = log_prefix
         self.process_runner = process_runner
         self.node_id = str(node_id)
@@ -374,7 +352,6 @@ class SSHOptions:
 class SSHCommandRunner(CommandRunnerInterface):
     def __init__(self, log_prefix, node_id, provider, auth_config,
                  cluster_name, process_runner, use_internal_ip):
-        super().__init__()
         ssh_control_hash = hashlib.md5(cluster_name.encode()).hexdigest()
         ssh_user_hash = hashlib.md5(getuser().encode()).hexdigest()
         ssh_control_path = "/tmp/ray_ssh_{}/{}".format(
@@ -402,7 +379,7 @@ class SSHCommandRunner(CommandRunnerInterface):
         else:
             return self.provider.external_ip(self.node_id)
 
-    def wait_for_ip(self, deadline):
+    def _wait_for_ip(self, deadline):
         # if we have IP do not print waiting info
         ip = self._get_node_ip()
         if ip is not None:
@@ -434,7 +411,7 @@ class SSHCommandRunner(CommandRunnerInterface):
         #   I think that's reasonable.
         deadline = time.time() + NODE_START_WAIT_S
         with LogTimer(self.log_prefix + "Got IP"):
-            ip = self.wait_for_ip(deadline)
+            ip = self._wait_for_ip(deadline)
 
             cli_logger.doassert(ip is not None,
                                 "Could not get node IP.")  # todo: msg
@@ -614,7 +591,6 @@ class SSHCommandRunner(CommandRunnerInterface):
 
 class DockerCommandRunner(CommandRunnerInterface):
     def __init__(self, docker_config, **common_args):
-        super().__init__()
         self.ssh_command_runner = SSHCommandRunner(**common_args)
         self.container_name = docker_config["container_name"]
         self.docker_config = docker_config
