+++ python/ray/tests/test_command_runner.py
@@ -1,7 +1,8 @@
 import pytest
 from ray.tests.test_autoscaler import MockProvider, MockProcessRunner
-from ray.autoscaler.command_runner import SSHCommandRunner, \
-    _with_environment_variables, DockerCommandRunner, KubernetesCommandRunner
+from ray.autoscaler.command_runner import CommandRunnerInterface, \
+    SSHCommandRunner, _with_environment_variables, DockerCommandRunner, \
+    KubernetesCommandRunner
 from ray.autoscaler.docker import DOCKER_MOUNT_PREFIX
 from getpass import getuser
 import hashlib
@@ -25,6 +26,31 @@ def test_environment_variable_encoder_dict():
 
     expected = """export value1='"string1"';export value2='{"a":"b","c":2}';echo hello"""  # noqa: E501
     assert res == expected
+
+
+def test_command_runner_interface_abstraction_violation():
+    """Enforces the CommandRunnerInterface functions on the subclasses.
+
+    This is important to make sure the subclasses do not violate the
+    function abstractions. If you need to add a new function to one of
+    the CommandRunnerInterface subclasses, you have to add it to
+    CommandRunnerInterface and all of its subclasses.
+    """
+
+    cmd_runner_interface_public_functions = dir(CommandRunnerInterface)
+    allowed_public_interface_functions = set([
+        func for func in cmd_runner_interface_public_functions
+        if not func.startswith("_")
+    ])
+    for subcls in [
+            SSHCommandRunner, DockerCommandRunner, KubernetesCommandRunner
+    ]:
+        subclass_available_functions = dir(subcls)
+        subclass_public_functions = set([
+            func for func in subclass_available_functions
+            if not func.startswith("_")
+        ])
+        assert allowed_public_interface_functions == subclass_public_functions
 
 
 def test_ssh_command_runner():
